rename to RPCClient and add getTransaction method
