+++ packages/backend/src/config/Config.ts
@@ -5,7 +5,7 @@ import { ChainId, ProjectId, Token, UnixTime } from '@l2beat/shared-pure'
 import { Knex } from 'knex'
 
 import { Project } from '../model'
-import { MulticallConfigEntry } from '../peripherals/ethereum/multicall/types'
+import { MulticallConfigEntry } from '../peripherals/multicall/types'
 
 export interface Config {
   readonly name: string
