+++ packages/backend/src/config/getChainTvlConfig.ts
@@ -2,7 +2,7 @@ import { Env } from '@l2beat/backend-tools'
 import { chains, layer2s } from '@l2beat/config'
 import { ChainId, ProjectId, UnixTime } from '@l2beat/shared-pure'
 
-import { toMulticallConfigEntry } from '../peripherals/ethereum/multicall/MulticallConfig'
+import { toMulticallConfigEntry } from '../peripherals/multicall/MulticallConfig'
 import { ChainTvlConfig } from './Config'
 
 const DEFAULT_RPC_CALLS_PER_MINUTE = 60
