+++ packages/backend/src/core/totalSupply/TotalSupplyProvider.ts
@@ -8,8 +8,8 @@ import {
 import { BigNumber, utils } from 'ethers'
 
 import { TotalSupplyRecord } from '../../peripherals/database/TotalSupplyRepository'
-import { MulticallClient } from '../../peripherals/ethereum/multicall/MulticallClient'
-import { MulticallRequest } from '../../peripherals/ethereum/multicall/types'
+import { MulticallClient } from '../../peripherals/multicall/MulticallClient'
+import { MulticallRequest } from '../../peripherals/multicall/types'
 
 export interface TotalSupplyQuery {
   assetId: AssetId
